+++ Source/MeshIO/imstkVegaMeshIO.cpp
@@ -142,7 +142,7 @@ VegaMeshIO::copyCells(std::shared_ptr<vega::VolumetricMesh> vegaMesh,
                       VecDataArray<int, dim>& cells)
 {
     typename VecDataArray<int, dim>::VecType cell;
-    for (size_t cellId = 0; cellId < vegaMesh->getNumElements(); ++cellId)
+    for (int cellId = 0; cellId < vegaMesh->getNumElements(); ++cellId)
     {
         for (int i = 0; i < vegaMesh->getNumElementVertices(); ++i)
         {
